refactor: extract useScrollLock and useEscapeKey hooks, deduplicate modal logic across 7 components
- Create useScrollLock hook — saves/restores previous overflow for nested modal safety
- Create useEscapeKey hook — composable escape key handler with active flag
- Refactor CareGuideDetail to use useFocusTrap + useScrollLock + useEscapeKey (removes inline focus trap)
- Refactor PlantDetailModal to use shared hooks (removes manual focus trap + scroll lock)
- Refactor PropagationForm to use useScrollLock + useEscapeKey
- Refactor PlantImageGallery to use useScrollLock
- Refactor QuickCareActions to use useScrollLock
- Refactor ConfirmDialog to use useScrollLock + useEscapeKey
- Refactor Modal to use useScrollLock + useEscapeKey

Net -134 lines: replaced ~219 lines of duplicated scroll lock, escape key, and
focus trap logic across 7 components with 47 lines in 2 reusable hooks.
